requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.4.0
enkanetwork.py==1.4.5
frozenlist==1.4.1
gevent==24.2.1
gevent-eventemitter==2.1
googlesearch-python==1.2.4
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.6.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
multidict==6.0.5
mysql-connector-python==8.4.0
packaging==24.1
pip-tools==7.4.1
protobuf==5.27.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic_core
PyNaCl==1.5.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rfc3986==2.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
steam==1.4.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.2
vdf==3.4
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
